Created Class ListGenerator and moved function ListOfProduct
