+++ FORA/FORA.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ListGenerator.cs" />
     <Compile Include="Products\Bread.cs" />
     <Compile Include="Buyer.cs" />
     <Compile Include="Products\Cake.cs" />
